requirements.txt
psycopg2
pyparsing
Babel==2.10.3
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-Moment==1.0.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
wtforms[email]
python-dateutil==2.8.2
python-slugify==6.1.2
flask_login